app/requirements.txt
telepot==12.7
python-dotenv==0.14.0
requests==2.24.0
